[feat]login page, app change to notespage
