+++ encrypted-notes-app/package.json
@@ -12,12 +12,14 @@
   "dependencies": {
     "@tiptap/react": "^2.5.5",
     "@tiptap/starter-kit": "^2.5.5",
+    "crypto-js": "^4.2.0",
     "normalize.css": "^8.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
+    "@types/crypto-js": "^4.2.2",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@types/uuid": "^10.0.0",
